+++ AoE2ScenarioRms/debug/apply_debug.py
@@ -1,6 +1,5 @@
 from typing import TYPE_CHECKING
 
-
 if TYPE_CHECKING:
     from AoE2ScenarioRms import AoE2ScenarioRms
 
@@ -11,6 +10,7 @@ class ApplyDebug:
     Base class for all 'debug appliers'
 
     """
+
     def __init__(self, rms: 'AoE2ScenarioRms') -> None:
         super().__init__()