+++ AoE2ScenarioRms/debug/apply_no_clutter.py
@@ -2,7 +2,6 @@ from typing import TYPE_CHECKING
 
 from AoE2ScenarioRms.debug.apply_debug import ApplyDebug
 
-
 if TYPE_CHECKING:
     from AoE2ScenarioRms import AoE2ScenarioRms
 
@@ -13,6 +12,7 @@ class ApplyNoClutter(ApplyDebug):
     Set the entire map to have no elevation or layered terrain for extra clarity
 
     """
+
     def __init__(self, rms: 'AoE2ScenarioRms') -> None:
         super().__init__(rms)