+++ AoE2ScenarioRms/util/locator.py
@@ -22,60 +22,42 @@ class Locator:
 
     """
 
-    def __init__(
-            self,
-            name: str,
-            amount: int,
-            grouping_method: GroupingMethod,
-            group_size: int | Tuple[int, int],
-            grid_map: 'GridMap',
-            loose_grouping_distance: int = -1,
-    ) -> None:
-        self.name: str = name
-        self.amount: int = amount
-        self.grouping_method: GroupingMethod = grouping_method
-        self.group_size: int | Tuple[int, int] = group_size
-        self.grid_map: 'GridMap' = grid_map
-        self.loose_grouping_distance: int = loose_grouping_distance
+    @staticmethod
+    def create_groups(config: 'CreateObjectConfig', grid_map: 'GridMap') -> List[List[Tile]]:
+        """
+        Create groups for the given create object config.
 
-        self.map_size = grid_map.map_size
+        Args:
+            config: The config object to be used for the groups
+            grid_map: The grid map to be taken into account when creating the groups
 
-        self.available_tiles: List[Tile] = grid_map.available_tiles(shuffle=True)
+        Returns:
+            A list of groups. Each group is also a list of tiles.
+        """
+        available_tiles = grid_map.available_tiles(shuffle=True)
 
-    @classmethod
-    def from_create_object_config(cls, create: 'CreateObjectConfig', grid_map: 'GridMap'):
-        return cls(
-            name=create.name,
-            amount=create.max_potential_group_count,
-            grouping_method=create.grouping,
-            group_size=create.number_of_objects,
-            grid_map=grid_map,
-            loose_grouping_distance=create.loose_grouping_distance,
-        )
+        amount = config.max_potential_group_count
+        name = config.name
 
-    def get_valid_tiles(self, name: str = None, amount: int = None) -> List[List[Tile]]:
         tiles: List[List[Tile]] = []
-        amount = amount or self.amount
-        name = name or self.name
-
-        if amount * 2 > len(self.available_tiles):
+        if amount * 2 > len(available_tiles):
             warn(f"For group '{name}', the amount of groups requested is really high. "
-                 f"({amount} groups compared to {len(self.available_tiles)} available tiles).\n"
+                 f"({amount} groups compared to {len(available_tiles)} available tiles).\n"
                  f"Consider lowering the max amount of necessary groups for '{name}'.", SpawnFailureWarning)
 
         # starting_tiles = self.find_random_locations(amount)
-        starting_tiles = self.available_tiles[:amount]
+        starting_tiles = available_tiles[:amount]
 
         failed_spawns = 0
         for starting_tile in starting_tiles:
-            min_, size = self.randomize_group_size(self.group_size)
+            min_, size = Locator.randomize_group_size(config.number_of_objects)
 
             group = [starting_tile]
             size -= 1
 
             failed_attempts = 0
             while size > 0 and failed_attempts < 50:
-                nearby_tile = self.find_nearby_tile(group)
+                nearby_tile = Locator.find_nearby_tile(grid_map, config.grouping, group, config.loose_grouping_distance)
 
                 if nearby_tile is not None:
                     group.append(nearby_tile)
@@ -94,79 +76,100 @@ class Locator:
 
         return tiles
 
-    def find_nearby_tile(self, group: List[Tile]) -> Tile | None:
-        if self.grouping_method == GroupingMethod.TIGHT:
-            return self.find_random_adjacent_tile(random.choice(group), group)
-        elif self.grouping_method == GroupingMethod.LOOSE:
-            return self.find_random_tile_within_range(group[0], self.loose_grouping_distance, group)
+    @staticmethod
+    def find_nearby_tile(
+            grid_map: GridMap,
+            method: GroupingMethod,
+            group: List[Tile],
+            loose_grouping_distance: int = -1
+    ) -> Tile | None:
+        """
+        Find a nearby tile with the given GridMap based on the method given.
+
+        Args:
+            grid_map: The gridmap to respect
+            method: The method to use for choosing a GroupingMethod
+            group: The current tiles belonging to this group (to not return a tile twice)
+            loose_grouping_distance: The distance loosely grouped units have to be apart (only used for LOOSE grouping)
+
+        Returns:
+            The tile found or ``None`` if no tile was found (doesn't mean that there is no tile at all)
+        """
+        if method == GroupingMethod.TIGHT:
+            return Locator.find_random_adjacent_tile(grid_map, random.choice(group), group)
+        elif method == GroupingMethod.LOOSE:
+            return Locator.find_random_tile_within_range(grid_map, group[0], loose_grouping_distance, group)
         else:
             return None
 
-    def find_random_adjacent_tile(self, tile: Tile, group: List[Tile]) -> Tile | None:
+    @staticmethod
+    def find_random_adjacent_tile(
+            grid_map: GridMap,
+            tile: Tile,
+            group: List[Tile]
+    ) -> Tile | None:
+        """
+        Find a random adjacent tile (not diagonal) around the given tile
+
+        Args:
+            grid_map: The gridmap to respect
+            tile: The tile to look around
+            group: The current tiles belonging to this group (to not return a tile twice)
+
+        Returns:
+            A random tile around the given tile, or if all are invalid, returns ``None``
+        """
         offset = random.randrange(0, 4)
 
         tiles = TileUtil.adjacent(tile)
         for i in range(4):
             new_tile = tiles[(i + offset) % 4]
-            if new_tile not in group and self.grid_map.is_valid(TileLevel.RES, new_tile):
+            if new_tile not in group and grid_map.is_valid(TileLevel.RES, new_tile):
                 return new_tile
         return None
 
-    def find_random_tile_within_range(self, tile: Tile, range_: int, group: List[Tile]) -> Tile | None:
-        for _ in range(10):
+    @staticmethod
+    def find_random_tile_within_range(
+            grid_map: GridMap,
+            tile: Tile,
+            range_: int,
+            group: List[Tile],
+            attempts: int = 10,
+    ) -> Tile | None:
+        """
+        Find a random tile within a given range
+
+        Args:
+            grid_map: The gridmap to respect
+            tile: The tile to look around
+            range_: The range in which to search for tiles
+            group: The current tiles belonging to this group (to not return a tile twice)
+            attempts: The amount of attempts to find a tile around a given tile
+
+        Returns:
+            A random tile around the given tile within a range. If all searches resulted in invalid tiles, None is
+            returned. This does not mean there's no tiles within the range.
+        """
+        # Todo: Change to random.choice() to make sure `None` means no tiles are valid
+        for _ in range(attempts):
             x_offset = random.randint(-range_, range_)
             y_offset = random.randint(-range_, range_)
 
             new_tile = Tile(tile.x + x_offset, tile.y + y_offset)
-            if new_tile not in group and self.grid_map.is_valid(TileLevel.RES, new_tile):
+            if new_tile not in group and grid_map.is_valid(TileLevel.RES, new_tile):
                 return new_tile
         return None
 
-    def find_random_locations(self, count: int) -> List[Tile]:
-        """
-        Find a list of random locations based on the grid map. Unused since shuffled list is now used.
-        Might be removed in the future.
-        """
-        tries = 0
-        max_tries = count * 100
-        spawns_left = count
-
-        locs = []
-        while spawns_left > 0:
-            tries += 1
-            loc = self.find_available_location()
-
-            locs.append(loc)
-            self.grid_map.temp(loc)
-            spawns_left -= 1
-
-            if tries > max_tries:
-                break
-
-        for loc in locs:
-            self.grid_map.reset(loc)
-
-        return locs
-
-    def find_available_location(self) -> Tile:
-        """
-        Find a list of random locations based on the grid map. Unused since shuffled list is now used.
-        Might be removed in the future.
-        """
-        counter = 1000
-        while counter > 0:
-            counter -= 1
-            x = random.randrange(0, self.map_size)
-            y = random.randrange(0, self.map_size)
-
-            if self.grid_map.is_available(x, y):
-                return Tile(x, y)
-
-        raise LocationNotFoundError("Unable to find a valid location. "
-                                    "Please verify that the map and settings allow enough room for all spawn attempts.")
-
     @staticmethod
     def randomize_group_size(size: int | Tuple[int, int]) -> Tuple[int, int]:
+        """
+        Args:
+            size: The size of a group
+
+        Returns:
+            A tuple with the minimum valid size for a group as first value, and if the given size is a tuple the second
+            value is a randomized value between the 2 numbers. If an int was given, the int is returned as second value.
+        """
         if isinstance(size, tuple):
             return size[0], random.randint(size[0], size[1])
         else: