+++ AoE2ScenarioRms/util/unit_util.py
@@ -9,6 +9,8 @@ from ordered_set import OrderedSet
 class UnitUtil:
     @staticmethod
     def get_tiles_around_object(unit: Unit, range_: int) -> OrderedSet[Tile]:
+        """Get tiles around an object within a given range"""
+        # Todo: change this function to use TileUtil....
         area = actions.new_area_object(unit._uuid)
         # Range around a unit on both sides including the center tile (hence the + 1)
         radius = range_ * 2 + 1