+++ examples/main.py
@@ -24,8 +24,8 @@ ScenarioUtil.clear(scenario, ObjectClear.ALL & ~ObjectClear.CLIFFS)
 # On top of that, all trees (forests) and cliffs are also blocked with a radius of 1 around them (default)
 grid_map = GridMapFactory.block(
     scenario=scenario,
-    terrain_marks=TerrainMark.water_beach(),
-    object_marks=ObjectMark.all(),
+    terrain_marks=TerrainMark.WATER_BEACH,
+    object_marks=ObjectMark.ALL,
 )
 # Throw in the create_object_config (Can be found in the other example file) to start calculating spawn positions
 # and add triggers and XS